+++ MegaMonster.Services.Payment/WebApi/Dto`s/PaymentRequestDto.cs
@@ -5,7 +5,7 @@ public class PaymentRequestDto
     public string OrderId { get; set; }
     public string UserName { get; set; }
     public string TicketType { get; set; }
-    public decimal Sum { get; set; }
+    public double Sum { get; set; }
     public int Count { get; set; }
     public string Action { get; set; }
 }
